+++ app/(afterLogin)/post/_hooks/usePostForm.ts
@@ -1,63 +1,27 @@
-import { useQueryClient, useMutation } from '@tanstack/react-query';
-import { useRouter } from 'next/navigation';
 import { useForm, SubmitHandler } from 'react-hook-form';
-import toast from 'react-hot-toast';
 
 import { usePreviewImage } from '@/app/_hooks/usePreviewImage';
-import useOnAuth from '@/app/_hooks/useOnAuth';
-import { usePostStore } from '@/app/_store/usePost';
 import { IPostInputs } from '@/app/_types/post';
-import { writePost } from '@/app/(afterLogin)/home/_services/writePost';
+import useWritePost from '@/app/(afterLogin)/post/_hooks/useWritePost';
 
 const usePostForm = () => {
-  const queryClient = useQueryClient();
-  const router = useRouter();
   const {
     register,
     handleSubmit,
     watch,
     formState: { errors, isSubmitting },
   } = useForm<IPostInputs>({ mode: 'onSubmit' });
-  const { user } = useOnAuth();
-  const { mode, parentPostId, reset } = usePostStore();
   const watchImage = watch('photoUrl');
   const { previewImage } = usePreviewImage({ watchImage });
+  const { writeFeed } = useWritePost();
 
-  const writeFeed = useMutation({
-    mutationFn: (data: IPostInputs) => writePost({ user, data, parentPostId }),
-    onSuccess: () => {
-      if (parentPostId) {
-        queryClient.invalidateQueries({
-          queryKey: ['post', parentPostId, 'comments'],
-        });
-        queryClient.invalidateQueries({
-          queryKey: ['post', parentPostId],
-        });
-        router.back();
-      } else {
-        queryClient.invalidateQueries({ queryKey: ['posts'] });
-        router.replace('/home');
-      }
-      toast.success('게시물이 작성되었습니다!');
-    },
-    onError: error => {
-      console.error('Error writing post:', error);
-      toast.error('게시물 작성 실패');
-    },
-    onSettled: () => {
-      reset();
-    },
-  });
-
-  const onSubmit: SubmitHandler<IPostInputs> = data => {
-    writeFeed.mutate(data);
+  const onSubmit: SubmitHandler<IPostInputs> = async data => {
+    await writeFeed.mutateAsync(data);
   };
 
   return {
     handleSubmit,
     onSubmit,
-    mode,
-    user,
     register,
     errors,
     previewImage,
